inpi_get_acte

retourne le json de reponse de la requete sur un acte

inpi_get_acte(id = "", simp = TRUE)

Arguments

id

l'identifiant a interroger

simp

simplifier le json ? default = TRUE

Valeur de retour

json non parse de la reponse a la requete

Examples

inpi_get_acte(id = "")
#>                    updatedAt     siren
#> 1  2025-09-21T16:20:40+02:00 991644733
#> 2  2025-09-21T15:59:59+02:00 991644444
#> 3  2025-09-21T15:57:07+02:00 991644428
#> 4  2025-09-21T15:41:53+02:00 845262047
#> 5  2025-09-21T15:31:15+02:00 991644022
#> 6  2025-09-21T15:16:56+02:00 991643826
#> 7  2025-09-21T14:53:19+02:00 991643511
#> 8  2025-09-21T14:47:15+02:00 991643438
#> 9  2025-09-21T14:40:16+02:00 991643370
#> 10 2025-09-21T13:03:22+02:00 991603291
#> 11 2025-09-21T13:03:22+02:00 991603291
#> 12 2025-09-21T13:03:22+02:00 991603291
#> 13 2025-09-21T13:03:20+02:00 977551555
#> 14 2025-09-21T13:03:20+02:00 977551555
#> 15 2025-09-21T13:03:16+02:00 991642331
#> 16 2025-09-21T13:03:11+02:00 991396995
#> 17 2025-09-21T13:03:11+02:00 991396995
#> 18 2025-09-21T13:03:11+02:00 991396995
#> 19 2025-09-21T13:03:10+02:00 991161522
#> 20 2025-09-21T13:03:06+02:00 489920926
#>                                 denomination  dateDepot
#> 1                              Aurore CARIOU 2025-09-21
#> 2                          BENEDICTE COLLOMP 2025-09-21
#> 3                               Ons Larguech 2025-09-21
#> 4                            MAXIME LECOMPTE 2025-09-21
#> 5                   Lilia Sekhleyan Kerobyan 2025-09-21
#> 6                             Farid nekkachi 2025-09-21
#> 7                            FLORIE BOUSQUIE 2025-09-21
#> 8                    Jean-Francois Herbommez 2025-09-21
#> 9                                Komal Tahir 2025-09-21
#> 10                               SPEED PARTS 2025-09-20
#> 11                               SPEED PARTS 2025-09-19
#> 12                               SPEED PARTS 2025-09-19
#> 13                               Mayminco O1 2025-09-18
#> 14                               Mayminco O1 2025-09-18
#> 15                         Christelle Grimal 2025-09-21
#> 16                                       MDY 2025-09-15
#> 17                                       MDY 2025-09-15
#> 18                                       MDY 2025-09-15
#> 19                                       LAP 2025-09-05
#> 20 SOCIETE CIVILE IMMOBILIERE LES MARMOUZETS 2025-09-17
#>                                                nomDocument confidentiality
#> 1               Attestation-d-avoir_informe-votre-conjoint          Public
#> 2                                            AUTORCONJOINT          Public
#> 3                              declaration_honneur_TIAN_LI          Public
#> 4                                     img20250919_10195387          Public
#> 5                                                  Image_1          Public
#> 6  Bail_d_habitation_meublee_12_Rue_Domremy_75013_V_Signee          Public
#> 7                         Declaration_d_option_du_conjoint          Public
#> 8                                  Carte_d_identite_Sophie          Public
#> 9                      MUHAMMAD_AJMAIR_TAHIR_accord_expres          Public
#> 10               Liste_des_souscripteurs_certifie_conforme          Public
#> 11   Certificat_de_depot_des_fonds_liste_des_souscripteurs          Public
#> 12                  Statuts_certifie_conforme_a_l_original          Public
#> 13    Mayminco_O1_-_Extrait_decision_du_President_RKP_docx          Public
#> 14                     Mayminco_O1_-_Statuts_post_RKP_docx          Public
#> 15                   GRIMAL_CHRISTELLE_CAPE_2025_08_signed          Public
#> 16                                 LISTE_DES_SOUSCRIPTEURS          Public
#> 17                     DEPOT_DE_CAPITAL_-_BANQUE_POPULAIRE          Public
#> 18                                    STATUTS_CONSTITUTIFS          Public
#> 19                               STATUTS_SIGNES_18_08_2025          Public
#> 20                                 CESSION3_LES_MARMOUZETS          Public
#>    deleted typeDocument
#> 1    FALSE        PJ_68
#> 2    FALSE        PJ_94
#> 3    FALSE        PJ_94
#> 4    FALSE        PJ_01
#> 5    FALSE        PJ_94
#> 6    FALSE        PJ_61
#> 7    FALSE        PJ_94
#> 8    FALSE        PJ_94
#> 9    FALSE        PJ_94
#> 10   FALSE        PJ_06
#> 11   FALSE        PJ_06
#> 12   FALSE        PJ_01
#> 13   FALSE        PJ_54
#> 14   FALSE        PJ_02
#> 15   FALSE        PJ_61
#> 16   FALSE       PJ_138
#> 17   FALSE       PJ_180
#> 18   FALSE        PJ_01
#> 19   FALSE        PJ_01
#> 20   FALSE       PJ_190
#>                                                                                                                                                  libelle
#> 1                                                                Exemplaire de l’accord exprès et de l’information préalable du conjoint (daté et signé)
#> 2                                                                                                         Déclaration d’option du conjoint collaborateur
#> 3                                                                                                         Déclaration d’option du conjoint collaborateur
#> 4                                                                                                                                      Copie des statuts
#> 5                                                                                                         Déclaration d’option du conjoint collaborateur
#> 6  Copie du contrat d’appui au projet d’entreprise pour la création ou la reprise d’une activité économique certifiée conforme par le représentant légal
#> 7                                                                                                         Déclaration d’option du conjoint collaborateur
#> 8                                                                                                         Déclaration d’option du conjoint collaborateur
#> 9                                                                                                         Déclaration d’option du conjoint collaborateur
#> 10                                         Certificat du dépositaire des fonds correspondant aux souscriptions avec en annexe la liste des souscripteurs
#> 11                                         Certificat du dépositaire des fonds correspondant aux souscriptions avec en annexe la liste des souscripteurs
#> 12                                                                                                                                     Copie des statuts
#> 13                                                  PV ayant décidé et constaté la modification enregistrée, certifié conforme par le représentant légal
#> 14                                                                                                                          Copie des statuts mis à jour
#> 15 Copie du contrat d’appui au projet d’entreprise pour la création ou la reprise d’une activité économique certifiée conforme par le représentant légal
#> 16                                                                                                                               Liste des souscripteurs
#> 17                                                                                                                        Attestation de dépôt des fonds
#> 18                                                                                                                                     Copie des statuts
#> 19                                                                                                                                     Copie des statuts
#> 20                                                        Acte sous seing privé de modification de la répartition des parts sociales (cession, donation)
#>          numNat                       id
#> 1  J00175449925 68d009b83997ccd4a50b81b0
#> 2  J00175448067 68d004dfb376132fdd015298
#> 3  J00175447838 68d00433572c41619a0308e7
#> 4  J00175446236 68d000a12c99a2ed20026c6f
#> 5  J00175445766 68cffe23e75fc9abf20828f3
#> 6  J00175444595 68cffac822964c13a00e7516
#> 7  J00175442664 68cff53f2dfa607cd2098737
#> 8  J00175442102 68cff3d3b40c7da8be0bba37
#> 9  J00175441575 68cff230c1347f62d304dc98
#> 10 J00175184001 68cfdb7ad8a5ab32b80e9ee2
#> 11 J00175184001 68cfdb7ad8a5ab32b80e9edd
#> 12 J00175184001 68cfdb7ad8a5ab32b80e9edb
#> 13 J00175064146 68cfdb78fc3d4dc289004a2a
#> 14 J00175064146 68cfdb78fc3d4dc289004a29
#> 15 J00175434760 68cfdb74deb787a17608ff37
#> 16 J00173888041 68cfdb6feb9e0ecb3a00e83d
#> 17 J00173888041 68cfdb6feb9e0ecb3a00e83c
#> 18 J00173888041 68cfdb6feb9e0ecb3a00e83a
#> 19 J00172509200 68cfdb6edc10270e8205f7f5
#> 20 J00173350729 68cfdb6acf56cbe45901cbd1